+++ src/app/views/shared/dynamic-library-examples/example-components/tree-views/tree-views.component.ts
@@ -7,11 +7,11 @@ import {
 } from '@folkehelseinstituttet/angular-components';
 
 // import { alleAtcKoderSomItems } from './alle_atckoder_som_items'; // use this for stress testing fhi-tree-view-checkbox or fhi-tree-view-radio with large dataset
-
+//
 @Component({
-    selector: 'app-tree-views',
-    templateUrl: './tree-views.component.html',
-    standalone: false
+  selector: 'app-tree-views',
+  templateUrl: './tree-views.component.html',
+  standalone: false,
 })
 export class TreeViewsComponent implements OnInit {
   @Input() itemId!: string;